requirements.txt
# Arabic-Audio-Preprocessing-and-Feature-Extraction Requirements

# Core audio processing
librosa>=0.10.0
soundfile>=0.12.0
noisereduce>=3.0.0
pydub>=0.25.1
torchaudio>=2.0.0

# Machine Learning
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0

# NeMo ASR for VAD
nemo-toolkit[asr]>=1.18.0

# Pyannote Audio for Speaker Diarization and Embeddings
pyannote.audio>=3.0.0

# Vector Database
pymilvus>=2.3.0

# Visualization
matplotlib>=3.6.0

# Utilities
tqdm>=4.65.0
uuid
json
warnings
glob
os
shutil
datetime